+++ src/components/playlist/Playlist.jsx
@@ -6,7 +6,7 @@ export default function Playlist() {
   return (
     <div id="playlist" className="box neutral">
       <TabBar>
-        <Tab to="/playlist/queueing" iconName="chevron-right" name="Queuing" />
+        <Tab to="/playlist/queuing" iconName="chevron-right" name="Queuing" />
         <Tab to="/playlist/played" iconName="chevron-left" name="Played" />
         <Tab
           to="/playlist/player-errors"
